Move endpoints to delete connections & VConnections to internal API
